upd averageSocVol
